// autogpt_platform/frontend/src/app/api/__generated__/models/agentDetailsResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { AgentDetailsResponseSessionId } from "./agentDetailsResponseSessionId";
import type { AgentDetails } from "./agentDetails";
import type { AgentDetailsResponseGraphId } from "./agentDetailsResponseGraphId";
import type { AgentDetailsResponseGraphVersion } from "./agentDetailsResponseGraphVersion";

/**
 * Response for get_details action.
 */
export interface AgentDetailsResponse {
  type?: ResponseType;
  message: string;
  session_id?: AgentDetailsResponseSessionId;
  agent: AgentDetails;
  user_authenticated?: boolean;
  graph_id?: AgentDetailsResponseGraphId;
  graph_version?: AgentDetailsResponseGraphVersion;
}